+++ Hollowborn/HollowbornDoomKnight/ADarkKnight.cs
@@ -7,7 +7,7 @@
 //cs_include Scripts/Hollowborn/HollowbornDoomKnight/CoreHollowbornDoomKnight.cs
 //cs_include Scripts/Good/BLoD/CoreBLOD.cs
 //cs_include Scripts/Evil/SDKA/CoreSDKA.cs
-//cs_include Scripts/Evil/NecroticSwordOfDoom.cs
+//cs_include Scripts/Evil/NSoD/CoreNSOD.cs
 //cs_include Scripts/Nation/CoreNation.cs
 //cs_include Scripts/CoreDailies.cs
 //cs_include Scripts/Chaos/DrakathsArmor.cs
